Isolate blackbox process platforms
