+++ CMakeLists.txt
@@ -1631,6 +1631,7 @@ if(GECODE_INSTALL)
       PATTERN "mznlib" EXCLUDE
       PATTERN "flatzinc/blackbox.hh" EXCLUDE
       PATTERN "flatzinc/blackbox-backend.hh" EXCLUDE
+      PATTERN "flatzinc/blackbox-process.hh" EXCLUDE
       PATTERN "exampleplugin" EXCLUDE
       PATTERN "standalone-example" EXCLUDE
       PATTERN "abi*" EXCLUDE)
