+++ Makefile.in
@@ -822,12 +822,13 @@ endif
 #
 
 FLATZINCSRC0 = flatzinc.cpp registry.cpp branch.cpp blackbox-backend.cpp \
-  blackbox-propagator.cpp
+  blackbox-process-none.cpp blackbox-process-posix.cpp \
+  blackbox-process-windows.cpp blackbox-propagator.cpp
 FLATZINC_GENSRC0 = parser.tab.cpp lexer.yy.cpp
 FLATZINCHDR0 = ast.hh conexpr.hh option.hh parser.hh \
 	plugin.hh registry.hh symboltable.hh varspec.hh \
 	branch.hh branch.hpp lastval.hh complete.hh
-FLATZINCPRIVATEHDR0 = blackbox.hh blackbox-backend.hh
+FLATZINCPRIVATEHDR0 = blackbox.hh blackbox-backend.hh blackbox-process.hh
 
 FLATZINCSRC = $(FLATZINCSRC0:%=gecode/flatzinc/%)
 FLATZINC_GENSRC = $(FLATZINC_GENSRC0:%=gecode/flatzinc/%)
