+++ gecode/flatzinc/blackbox-backend.cpp
@@ -3,6 +3,9 @@
  *  Main authors:
  *     Jip J. Dekker <jip.dekker@monash.edu>
  *
+ *  Contributing authors:
+ *     Mikael Zayenz Lagerkvist <lagerkvist@gecode.dev>
+ *
  *  Copyright:
  *     Jip J. Dekker, 2026
  *
@@ -48,6 +51,7 @@
 #include <gecode/flatzinc.hh>
 #include <gecode/flatzinc/blackbox.hh>
 #include <gecode/flatzinc/blackbox-backend.hh>
+#include <gecode/flatzinc/blackbox-process.hh>
 
 #include <atomic>
 #include <cassert>
@@ -72,18 +76,6 @@
 #include <windows.h>
 #else
 #include <dlfcn.h>
-#ifdef GECODE_HAS_POSIX_BLACKBOX_EXEC
-#include <fcntl.h>
-#include <pthread.h>
-#include <signal.h>
-#include <spawn.h>
-#include <sys/socket.h>
-#include <sys/types.h>
-#include <sys/wait.h>
-#include <time.h>
-#include <unistd.h>
-extern char **environ;
-#endif
 #endif
 
 #ifdef GECODE_HAS_THREADS
@@ -120,80 +112,6 @@ windows_error(const std::string &prefix, DWORD err) {
   return prefix + " (Windows error " + std::to_string(err) + ")";
 }
 
-class WindowsHandle {
-private:
-  HANDLE handle;
-public:
-  explicit WindowsHandle(HANDLE handle0=NULL) : handle(handle0) {}
-  ~WindowsHandle(void) { reset(); }
-
-  WindowsHandle(const WindowsHandle &) = delete;
-  WindowsHandle &operator=(const WindowsHandle &) = delete;
-
-  HANDLE get(void) const { return handle; }
-  HANDLE *put(void) {
-    reset();
-    return &handle;
-  }
-  HANDLE release(void) {
-    HANDLE handle0 = handle;
-    handle = NULL;
-    return handle0;
-  }
-  bool valid(void) const {
-    return (handle != NULL) && (handle != INVALID_HANDLE_VALUE);
-  }
-  void reset(HANDLE handle0=NULL) {
-    if (valid()) {
-      CloseHandle(handle);
-    }
-    handle = handle0;
-  }
-};
-
-class WindowsAttributeList {
-private:
-  std::vector<char> buffer;
-  LPPROC_THREAD_ATTRIBUTE_LIST list;
-  bool initialized;
-public:
-  WindowsAttributeList(void) : list(NULL), initialized(false) {}
-  ~WindowsAttributeList(void) {
-    if (initialized) {
-      DeleteProcThreadAttributeList(list);
-    }
-  }
-
-  void init(void) {
-    SIZE_T size = 0;
-    InitializeProcThreadAttributeList(NULL, 1, 0, &size);
-    if (size == 0) {
-      throw Error("BlackBoxExec",
-                  windows_error("ProcThreadAttributeList size query failed",
-                                GetLastError()));
-    }
-    buffer.resize(size);
-    list = reinterpret_cast<LPPROC_THREAD_ATTRIBUTE_LIST>(buffer.data());
-    if (!InitializeProcThreadAttributeList(list, 1, 0, &size)) {
-      throw Error("BlackBoxExec",
-                  windows_error("InitializeProcThreadAttributeList failed",
-                                GetLastError()));
-    }
-    initialized = true;
-  }
-
-  void set_inherited_handles(HANDLE *handles, DWORD count) {
-    if (!UpdateProcThreadAttribute(list, 0, PROC_THREAD_ATTRIBUTE_HANDLE_LIST,
-                                   handles, sizeof(HANDLE) * count, NULL,
-                                   NULL)) {
-      throw Error("BlackBoxExec",
-                  windows_error("PROC_THREAD_ATTRIBUTE_HANDLE_LIST failed",
-                                GetLastError()));
-    }
-  }
-
-  LPPROC_THREAD_ATTRIBUTE_LIST get(void) const { return list; }
-};
 
 bool
 has_dll_suffix(const std::string &name) {
@@ -233,11 +151,6 @@ dll_candidates(const std::string &name) {
   return candidates;
 }
 
-bool
-qualified_path(const std::wstring &program) {
-  return (program.find_first_of(L"\\/") != std::wstring::npos) ||
-         ((program.size() > 1) && (program[1] == L':'));
-}
 
 void
 close_library(void *library) {
@@ -253,185 +166,8 @@ close_library(void *library) {
   }
 }
 
-#ifdef GECODE_HAS_POSIX_BLACKBOX_EXEC
-int
-set_cloexec(int fd) {
-  int flags = fcntl(fd, F_GETFD);
-  if (flags == -1) {
-    return -1;
-  }
-  return fcntl(fd, F_SETFD, flags | FD_CLOEXEC);
-}
-
-int
-dup_cloexec(int fd, int min_fd) {
-  int nfd;
-#ifdef F_DUPFD_CLOEXEC
-  nfd = fcntl(fd, F_DUPFD_CLOEXEC, min_fd);
-  if (nfd != -1) {
-    return nfd;
-  }
-  if (errno != EINVAL) {
-    return -1;
-  }
 #endif
-  nfd = fcntl(fd, F_DUPFD, min_fd);
-  if (nfd == -1) {
-    return -1;
-  }
-  if (set_cloexec(nfd) != 0) {
-    int e = errno;
-    ::close(nfd);
-    errno = e;
-    return -1;
-  }
-  return nfd;
-}
 
-int
-move_from_standard_fd(int fd) {
-  if (fd > STDERR_FILENO) {
-    return fd;
-  }
-  int nfd = dup_cloexec(fd, STDERR_FILENO + 1);
-  if (nfd == -1) {
-    return -1;
-  }
-  ::close(fd);
-  return nfd;
-}
-
-class FileDescriptor {
-private:
-  int fd;
-public:
-  explicit FileDescriptor(int fd0=-1) : fd(fd0) {}
-  ~FileDescriptor(void) { reset(); }
-
-  int get(void) const { return fd; }
-  int release(void) {
-    int fd0 = fd;
-    fd = -1;
-    return fd0;
-  }
-  void reset(int fd0=-1) {
-    if (fd != -1) {
-      ::close(fd);
-    }
-    fd = fd0;
-  }
-};
-
-int
-move_away_from_standard_fd(FileDescriptor &fd) {
-  int old = fd.release();
-  int nfd = move_from_standard_fd(old);
-  if (nfd == -1) {
-    fd.reset(old);
-  } else {
-    fd.reset(nfd);
-  }
-  return nfd;
-}
-
-class SpawnFileActions {
-private:
-  posix_spawn_file_actions_t actions;
-  bool initialized;
-public:
-  SpawnFileActions(void) : initialized(false) {}
-  ~SpawnFileActions(void) {
-    if (initialized) {
-      posix_spawn_file_actions_destroy(&actions);
-    }
-  }
-
-  int init(void) {
-    int err = posix_spawn_file_actions_init(&actions);
-    initialized = err == 0;
-    return err;
-  }
-  posix_spawn_file_actions_t *get(void) { return &actions; }
-};
-
-class SpawnAttributes {
-private:
-  posix_spawnattr_t attr;
-  bool initialized;
-public:
-  SpawnAttributes(void) : initialized(false) {}
-  ~SpawnAttributes(void) {
-    if (initialized) {
-      posix_spawnattr_destroy(&attr);
-    }
-  }
-
-  int init(void) {
-    int err = posix_spawnattr_init(&attr);
-    initialized = err == 0;
-    return err;
-  }
-  posix_spawnattr_t *get(void) { return &attr; }
-};
-
-int
-create_socketpair(int sv[2]) {
-#ifdef SOCK_CLOEXEC
-  if (socketpair(AF_UNIX, SOCK_STREAM | SOCK_CLOEXEC, 0, sv) == 0) {
-    return 0;
-  }
-  if (errno != EINVAL) {
-    return -1;
-  }
-#endif
-  if (socketpair(AF_UNIX, SOCK_STREAM, 0, sv) != 0) {
-    return -1;
-  }
-  if ((set_cloexec(sv[0]) != 0) || (set_cloexec(sv[1]) != 0)) {
-    int e = errno;
-    ::close(sv[0]);
-    ::close(sv[1]);
-    errno = e;
-    return -1;
-  }
-  return 0;
-}
-
-ssize_t
-send_no_sigpipe(int fd, const char *data, size_t size) {
-#ifdef MSG_NOSIGNAL
-  return send(fd, data, size, MSG_NOSIGNAL);
-#else
-#ifdef SO_NOSIGPIPE
-  return send(fd, data, size, 0);
-#else
-  sigset_t block;
-  sigset_t old;
-  sigset_t pending;
-  sigemptyset(&block);
-  sigaddset(&block, SIGPIPE);
-  bool blocked = false;
-  bool was_pending = false;
-  if (pthread_sigmask(SIG_BLOCK, &block, &old) == 0) {
-    blocked = true;
-    if (sigpending(&pending) == 0) {
-      was_pending = sigismember(&pending, SIGPIPE) == 1;
-    }
-  }
-  ssize_t n = send(fd, data, size, 0);
-  if ((n == -1) && (errno == EPIPE) && !was_pending) {
-    const struct timespec timeout = {0, 0};
-    sigtimedwait(&block, NULL, &timeout);
-  }
-  if (blocked) {
-    pthread_sigmask(SIG_SETMASK, &old, NULL);
-  }
-  return n;
-#endif
-#endif
-}
-#endif
-#endif
 
 template<class T>
 T
@@ -500,8 +236,6 @@ check_floats(const std::vector<double> &v, const char *source) {
   }
 }
 #endif
-
-const size_t max_exec_response_size = 1024 * 1024;
 
 } // namespace
 
@@ -701,603 +435,6 @@ BlackBoxLibrary::run(BlackBoxCall& call) {
 #endif
 }
 
-class BlackBoxExec::Session {
-protected:
-#ifdef _WIN32
-  HANDLE job;
-  HANDLE process;
-  HANDLE pipe_send;
-  HANDLE pipe_receive;
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-  pid_t child;
-  int pipe_send;
-  FILE *file_receive;
-#endif
-#ifdef GECODE_HAS_THREADS
-  std::thread::id owner;
-#endif
-
-  static std::string last_error(const std::string &prefix) {
-#ifdef _WIN32
-    return prefix + " (Windows error " + std::to_string(GetLastError()) + ")";
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-    return prefix + " (errno " + std::to_string(errno) + ")";
-#else
-    return prefix;
-#endif
-  }
-
-#ifdef _WIN32
-  static void close_handle(HANDLE &h) {
-    if (h != NULL) {
-      CloseHandle(h);
-      h = NULL;
-    }
-  }
-
-  static std::wstring quote_argument(const std::wstring &arg) {
-    std::wstring q(L"\"");
-    unsigned int backslashes = 0;
-    for (wchar_t ch : arg) {
-      if (ch == L'\\') {
-        backslashes++;
-      } else if (ch == L'"') {
-        q.append(backslashes * 2 + 1, L'\\');
-        q += ch;
-        backslashes = 0;
-      } else {
-        q.append(backslashes, L'\\');
-        q += ch;
-        backslashes = 0;
-      }
-    }
-    q.append(backslashes * 2, L'\\');
-    q += L'"';
-    return q;
-  }
-
-  void open_windows(const std::string &program,
-                    const std::vector<std::string> &args);
-  void close_windows(void);
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-  static void sleep_grace_period(void) {
-    struct timespec remaining = {0, 10000000};
-    while ((nanosleep(&remaining, &remaining) == -1) && (errno == EINTR)) {}
-  }
-
-  static bool child_exited(pid_t pid) {
-    siginfo_t info;
-    do {
-      info.si_pid = 0;
-      if (waitid(P_PID, pid, &info, WEXITED | WNOHANG | WNOWAIT) == 0) {
-        return info.si_pid != 0;
-      }
-    } while (errno == EINTR);
-    return false;
-  }
-
-  static void signal_group(pid_t pid, int signal) {
-    if ((kill(-pid, signal) == -1) && (errno == ESRCH)) {
-      return;
-    }
-  }
-
-  static void wait_group(pid_t pid, int attempts) {
-    for (int i = 0; i < attempts; i++) {
-      if ((kill(-pid, 0) == -1) && (errno == ESRCH)) {
-        return;
-      }
-      if (child_exited(pid)) {
-        return;
-      }
-      sleep_grace_period();
-    }
-  }
-
-  static void terminate_child(pid_t pid) {
-    if (pid <= 0) {
-      return;
-    }
-    int status = 0;
-    // Keep the child unreaped until the group has received both signals.
-    signal_group(pid, SIGTERM);
-    wait_group(pid, 100);
-    signal_group(pid, SIGKILL);
-    do {
-      if (waitpid(pid, &status, 0) != -1) {
-        return;
-      }
-    } while (errno == EINTR);
-  }
-
-  static void check_sigchld(void) {
-    struct sigaction action;
-    if (sigaction(SIGCHLD, NULL, &action) != 0) {
-      throw Error("BlackBoxExec", last_error("SIGCHLD query failed"));
-    }
-    if ((action.sa_handler != SIG_DFL)
-#ifdef SA_NOCLDWAIT
-        || (action.sa_flags & SA_NOCLDWAIT)
-#endif
-       ) {
-      throw Error("BlackBoxExec",
-                  "Cannot start a blackbox process unless SIGCHLD uses "
-                  "SIG_DFL without SA_NOCLDWAIT");
-    }
-  }
-
-  void open_posix(const std::string &program,
-                  const std::vector<std::string> &args);
-  void close_posix(void);
-#endif
-
-public:
-  Session(const std::string &program, const std::vector<std::string> &args)
-#ifdef _WIN32
-      : job(NULL), process(NULL), pipe_send(NULL), pipe_receive(NULL)
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-      : child(-1), pipe_send(-1), file_receive(NULL)
-#endif
-  {
-#ifdef GECODE_HAS_THREADS
-    owner = std::this_thread::get_id();
-#endif
-#ifdef _WIN32
-    open_windows(program, args);
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-    open_posix(program, args);
-#else
-    (void)program;
-    (void)args;
-    throw Error("BlackBoxExec",
-                "Persistent process blackboxes are not supported on this "
-                "platform");
-#endif
-  }
-
-  ~Session(void) { close(); }
-
-  bool owned_by_current_thread(void) const {
-#ifdef GECODE_HAS_THREADS
-    return owner == std::this_thread::get_id();
-#else
-    return true;
-#endif
-  }
-
-  std::string run(const std::string &out_buf) {
-#ifdef _WIN32
-    size_t written = 0;
-    while (written < out_buf.size()) {
-      DWORD count = 0;
-      DWORD remaining =
-          static_cast<DWORD>(out_buf.size() - written);
-      BOOL success =
-          WriteFile(pipe_send, out_buf.data() + written, remaining, &count,
-                    nullptr);
-      if (!success || count == 0) {
-        throw Error("BlackBoxExec",
-                    last_error("Writing blackbox process input failed"));
-      }
-      written += count;
-    }
-
-    char c[2] = {0, 0};
-    std::ostringstream oss;
-    size_t response_size = 0;
-    while (c[0] != '\n') {
-      DWORD count = 0;
-      BOOL success = ReadFile(pipe_receive, c, sizeof(c) - 1, &count, NULL);
-      if (!success) {
-        if (GetLastError() == ERROR_BROKEN_PIPE) {
-          throw Error("BlackBoxExec",
-                      "Blackbox process provided an incomplete response");
-        }
-        throw Error(
-            "BlackBoxExec",
-            "Failed to read blackbox process output from pipe");
-      } else if (count == 0) {
-        throw Error("BlackBoxExec",
-                    "Blackbox process provided an incomplete response");
-      }
-      assert(count == 1);
-      if (++response_size > max_exec_response_size) {
-        throw Error("BlackBoxExec",
-                    "Blackbox process response exceeds the size limit");
-      }
-      oss << c[0];
-    }
-    return oss.str();
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-    const char *p = out_buf.c_str();
-    size_t remaining = out_buf.size();
-    while (remaining > 0) {
-      ssize_t n = send_no_sigpipe(pipe_send, p, remaining);
-      if (n < 0) {
-        if (errno == EINTR) {
-          continue;
-        }
-        throw Error("BlackBoxExec",
-                    "Writing blackbox process input failed with errno " +
-                        std::to_string(errno));
-      }
-      if (n == 0) {
-        throw Error("BlackBoxExec",
-                    "Writing blackbox process input wrote zero bytes");
-      }
-      p += n;
-      remaining -= static_cast<size_t>(n);
-    }
-
-    std::string in_buffer;
-    while (true) {
-      errno = 0;
-      int ch = fgetc(file_receive);
-      if (ch == EOF) {
-        if (feof(file_receive)) {
-          throw Error("BlackBoxExec",
-                      "Blackbox process provided an incomplete response");
-        }
-        int err = errno;
-        if (err == EINTR) {
-          clearerr(file_receive);
-          continue;
-        }
-        throw Error("BlackBoxExec",
-                    std::string("Reading blackbox process output from pipe "
-                                "failed with errno ") +
-                        std::to_string(err));
-      }
-      in_buffer += static_cast<char>(ch);
-      if (in_buffer.size() > max_exec_response_size) {
-        throw Error("BlackBoxExec",
-                    "Blackbox process response exceeds the size limit");
-      }
-      if (ch == '\n') {
-        break;
-      }
-    }
-    return in_buffer;
-#else
-    (void)out_buf;
-    throw Error("BlackBoxExec",
-                "Persistent process blackboxes are not supported on this "
-                "platform");
-#endif
-  }
-
-  void close(void) {
-#ifdef _WIN32
-    close_windows();
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-    close_posix();
-#endif
-  }
-};
-
-#ifdef _WIN32
-void
-BlackBoxExec::Session::open_windows(const std::string &program,
-                                    const std::vector<std::string> &args) {
-  // Build the command line before opening OS handles so allocation/conversion
-  // failures cannot leak partially constructed process state.
-  std::wstring program_w = utf8_to_wide(program);
-  std::wstring prog = quote_argument(program_w);
-  for (const std::string &a : args) {
-    prog += L" ";
-    prog += quote_argument(utf8_to_wide(a));
-  }
-  std::vector<wchar_t> cmdline(prog.begin(), prog.end());
-  cmdline.push_back(L'\0');
-
-  SECURITY_ATTRIBUTES saAttr;
-  saAttr.nLength = sizeof(SECURITY_ATTRIBUTES);
-  saAttr.bInheritHandle = TRUE;
-  saAttr.lpSecurityDescriptor = NULL;
-
-  WindowsHandle child_stdin_read;
-  WindowsHandle child_stdin_write;
-  WindowsHandle child_stdout_read;
-  WindowsHandle child_stdout_write;
-  WindowsHandle child_stderr_write;
-  if (!CreatePipe(child_stdout_read.put(), child_stdout_write.put(), &saAttr,
-                  0)) {
-    throw Error("BlackBoxExec", last_error("Stdout CreatePipe failed"));
-  }
-  if (!SetHandleInformation(child_stdout_read.get(), HANDLE_FLAG_INHERIT, 0)) {
-    throw Error("BlackBoxExec",
-                last_error("Stdout SetHandleInformation failed"));
-  }
-  if (!CreatePipe(child_stdin_read.put(), child_stdin_write.put(), &saAttr,
-                  0)) {
-    throw Error("BlackBoxExec", last_error("Stdin CreatePipe failed"));
-  }
-  if (!SetHandleInformation(child_stdin_write.get(), HANDLE_FLAG_INHERIT, 0)) {
-    throw Error("BlackBoxExec",
-                last_error("Stdin SetHandleInformation failed"));
-  }
-
-  HANDLE parent_stderr = GetStdHandle(STD_ERROR_HANDLE);
-  if ((parent_stderr != NULL) && (parent_stderr != INVALID_HANDLE_VALUE)) {
-    if (!DuplicateHandle(GetCurrentProcess(), parent_stderr,
-                         GetCurrentProcess(), child_stderr_write.put(), 0, TRUE,
-                         DUPLICATE_SAME_ACCESS)) {
-      throw Error("BlackBoxExec",
-                  last_error("stderr DuplicateHandle failed"));
-    }
-  } else {
-    HANDLE nul = CreateFileW(L"NUL", GENERIC_WRITE,
-                             FILE_SHARE_READ | FILE_SHARE_WRITE, &saAttr,
-                             OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
-    if (nul == INVALID_HANDLE_VALUE) {
-      throw Error("BlackBoxExec", last_error("stderr NUL CreateFile failed"));
-    }
-    child_stderr_write.reset(nul);
-  }
-
-  WindowsAttributeList attr_list;
-  attr_list.init();
-  PROCESS_INFORMATION piProcInfo;
-  STARTUPINFOEXW siStartInfo;
-  ZeroMemory(&piProcInfo, sizeof(PROCESS_INFORMATION));
-  ZeroMemory(&siStartInfo, sizeof(STARTUPINFOEXW));
-  siStartInfo.StartupInfo.cb = sizeof(STARTUPINFOEXW);
-  siStartInfo.StartupInfo.hStdOutput = child_stdout_write.get();
-  siStartInfo.StartupInfo.hStdInput = child_stdin_read.get();
-  siStartInfo.StartupInfo.hStdError = child_stderr_write.get();
-  siStartInfo.StartupInfo.dwFlags |= STARTF_USESTDHANDLES;
-
-  HANDLE inherit_handles[3] = {child_stdin_read.get(), child_stdout_write.get(),
-                               child_stderr_write.get()};
-  attr_list.set_inherited_handles(inherit_handles, 3);
-  siStartInfo.lpAttributeList = attr_list.get();
-
-  WindowsHandle process_job(CreateJobObjectW(NULL, NULL));
-  if (!process_job.valid()) {
-    throw Error("BlackBoxExec", last_error("CreateJobObject failed"));
-  }
-  JOBOBJECT_EXTENDED_LIMIT_INFORMATION job_info;
-  ZeroMemory(&job_info, sizeof(job_info));
-  job_info.BasicLimitInformation.LimitFlags = JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE;
-  if (!SetInformationJobObject(process_job.get(),
-                               JobObjectExtendedLimitInformation, &job_info,
-                               sizeof(job_info))) {
-    throw Error("BlackBoxExec", last_error("SetInformationJobObject failed"));
-  }
-
-  BOOL processStarted =
-      CreateProcessW(qualified_path(program_w) ? program_w.c_str() : NULL,
-                     cmdline.data(),        // command line
-                     nullptr,               // process security attributes
-                     nullptr,               // primary thread security attributes
-                     TRUE,                  // handles from attribute list
-                     EXTENDED_STARTUPINFO_PRESENT | CREATE_SUSPENDED,
-                     nullptr,               // use parent's environment
-                     nullptr,               // use parent's current directory
-                     &siStartInfo.StartupInfo,
-                     &piProcInfo);          // receives PROCESS_INFORMATION
-
-  if (!processStarted) {
-    throw Error("BlackBoxExec",
-                windows_error("starting blackbox process failed for program `" +
-                              program + "'", GetLastError()));
-  }
-  WindowsHandle process_handle(piProcInfo.hProcess);
-  WindowsHandle thread_handle(piProcInfo.hThread);
-  if (!AssignProcessToJobObject(process_job.get(), process_handle.get())) {
-    DWORD err = GetLastError();
-    DWORD terminate_err = ERROR_SUCCESS;
-    if (!TerminateProcess(process_handle.get(), 1)) {
-      terminate_err = GetLastError();
-    }
-    DWORD wait = WaitForSingleObject(process_handle.get(), 5000);
-    std::string message = windows_error(
-        "Unable to assign blackbox process to required job", err);
-    if (terminate_err != ERROR_SUCCESS) {
-      message += "; " + windows_error("TerminateProcess cleanup failed",
-                                       terminate_err);
-    }
-    if (wait == WAIT_FAILED) {
-      message += "; " + last_error("process cleanup wait failed");
-    } else if (wait == WAIT_TIMEOUT) {
-      message += "; process cleanup timed out";
-    }
-    throw Error("BlackBoxExec", message);
-  }
-
-  if (ResumeThread(thread_handle.get()) == static_cast<DWORD>(-1)) {
-    DWORD err = GetLastError();
-    DWORD terminate_err = ERROR_SUCCESS;
-    if (!TerminateJobObject(process_job.get(), 1)) {
-      terminate_err = GetLastError();
-    }
-    HANDLE assigned_job = process_job.release();
-    DWORD close_err = ERROR_SUCCESS;
-    if (!CloseHandle(assigned_job)) {
-      close_err = GetLastError();
-    }
-    DWORD wait = WaitForSingleObject(process_handle.get(), 5000);
-    std::string message = windows_error(
-        "ResumeThread failed for blackbox process", err);
-    if (terminate_err != ERROR_SUCCESS) {
-      message += "; " + windows_error("TerminateJobObject cleanup failed",
-                                       terminate_err);
-    }
-    if (close_err != ERROR_SUCCESS) {
-      message += "; " + windows_error("job cleanup close failed", close_err);
-    }
-    if (wait == WAIT_FAILED) {
-      message += "; " + last_error("process cleanup wait failed");
-    } else if (wait == WAIT_TIMEOUT) {
-      message += "; process cleanup timed out";
-    }
-    throw Error("BlackBoxExec", message);
-  }
-
-  pipe_send = child_stdin_write.release();
-  pipe_receive = child_stdout_read.release();
-  process = process_handle.release();
-  job = process_job.release();
-}
-
-void
-BlackBoxExec::Session::close_windows(void) {
-  close_handle(pipe_send);
-  close_handle(pipe_receive);
-  if (process != NULL) {
-    DWORD wait = WaitForSingleObject(process, 1000);
-    if (wait == WAIT_TIMEOUT) {
-      if (job != NULL) {
-        TerminateJobObject(job, 1);
-      } else {
-        TerminateProcess(process, 1);
-      }
-      WaitForSingleObject(process, 5000);
-    }
-    close_handle(process);
-  }
-  close_handle(job);
-}
-#elif defined(GECODE_HAS_POSIX_BLACKBOX_EXEC)
-void
-BlackBoxExec::Session::open_posix(const std::string &program,
-                                  const std::vector<std::string> &args) {
-  const int READ = 0;
-  const int WRITE = 1;
-
-  std::vector<char *> argv;
-  argv.reserve(args.size() + 2);
-  argv.push_back(const_cast<char *>(program.c_str()));
-  for (const std::string &a : args) {
-    argv.push_back(const_cast<char *>(a.c_str()));
-  }
-  argv.push_back(nullptr);
-
-  check_sigchld();
-
-  FileDescriptor child_in[2];
-  FileDescriptor child_out[2];
-  int fds[2];
-  if (create_socketpair(fds) != 0) {
-    throw Error("BlackBoxExec", last_error("stdin socket creation failed"));
-  }
-  child_in[READ].reset(fds[READ]);
-  child_in[WRITE].reset(fds[WRITE]);
-  if (create_socketpair(fds) != 0) {
-    throw Error("BlackBoxExec", last_error("stdout socket creation failed"));
-  }
-  child_out[READ].reset(fds[READ]);
-  child_out[WRITE].reset(fds[WRITE]);
-  FileDescriptor *session_fds[] = {
-    &child_in[READ], &child_in[WRITE],
-    &child_out[READ], &child_out[WRITE]
-  };
-  for (FileDescriptor *fd : session_fds) {
-    if (move_away_from_standard_fd(*fd) == -1) {
-      throw Error("BlackBoxExec",
-                  last_error("moving session descriptors away from stdio "
-                             "failed"));
-    }
-  }
-
-  SpawnFileActions actions;
-  int err = actions.init();
-  if (err != 0) {
-    errno = err;
-    throw Error("BlackBoxExec", last_error("spawn file action init failed"));
-  }
-
-  SpawnAttributes attr;
-  err = attr.init();
-  if (err != 0) {
-    errno = err;
-    throw Error("BlackBoxExec", last_error("spawn attribute init failed"));
-  }
-
-  err = posix_spawnattr_setpgroup(attr.get(), 0);
-  if (err == 0) {
-    short flags = POSIX_SPAWN_SETPGROUP;
-#if defined(GECODE_HAS_POSIX_SPAWN_CLOEXEC_DEFAULT) && \
-    defined(GECODE_HAS_POSIX_SPAWN_ADDINHERIT_NP)
-    flags |= POSIX_SPAWN_CLOEXEC_DEFAULT;
-#endif
-    err = posix_spawnattr_setflags(attr.get(), flags);
-  }
-  if (err == 0) {
-    err = posix_spawn_file_actions_adddup2(actions.get(),
-                                           child_in[READ].get(),
-                                           STDIN_FILENO);
-  }
-  if (err == 0) {
-    err = posix_spawn_file_actions_adddup2(actions.get(),
-                                           child_out[WRITE].get(),
-                                           STDOUT_FILENO);
-  }
-#if defined(GECODE_HAS_POSIX_SPAWN_CLOEXEC_DEFAULT) && \
-    defined(GECODE_HAS_POSIX_SPAWN_ADDINHERIT_NP)
-  if (err == 0) {
-    err = posix_spawn_file_actions_addinherit_np(actions.get(),
-                                                  STDERR_FILENO);
-  }
-#elif defined(GECODE_HAS_POSIX_SPAWN_ADDCLOSEFROM_NP)
-  if (err == 0) {
-    err = posix_spawn_file_actions_addclosefrom_np(actions.get(),
-                                                    STDERR_FILENO + 1);
-  }
-#endif
-  if (err == 0) {
-    err = posix_spawnp(&child, program.c_str(), actions.get(), attr.get(),
-                       argv.data(), environ);
-  }
-  if (err != 0) {
-    child = -1;
-    errno = err;
-    throw Error("BlackBoxExec", last_error("starting blackbox process failed"));
-  }
-
-  child_in[READ].reset();
-  child_out[WRITE].reset();
-
-#ifdef SO_NOSIGPIPE
-  int nosigpipe = 1;
-  if (setsockopt(child_in[WRITE].get(), SOL_SOCKET, SO_NOSIGPIPE, &nosigpipe,
-                 sizeof(nosigpipe)) != 0) {
-    int e = errno;
-    terminate_child(child);
-    child = -1;
-    errno = e;
-    throw Error("BlackBoxExec", last_error("SO_NOSIGPIPE setup failed"));
-  }
-#endif
-  FILE *receive = fdopen(child_out[READ].get(), "r");
-  if (receive == NULL) {
-    int e = errno;
-    terminate_child(child);
-    child = -1;
-    errno = e;
-    throw Error("BlackBoxExec", last_error("fdopen failed"));
-  }
-  file_receive = receive;
-  child_out[READ].release();
-  pipe_send = child_in[WRITE].release();
-}
-
-void
-BlackBoxExec::Session::close_posix(void) {
-  if (pipe_send != -1) {
-    ::close(pipe_send);
-    pipe_send = -1;
-  }
-  if (file_receive != NULL) {
-    fclose(file_receive);
-    file_receive = NULL;
-  }
-  if (child > 0) {
-    terminate_child(child);
-    child = -1;
-  }
-}
-#endif
 
 BlackBoxExec::BlackBoxExec(const std::string &program0,
                            const std::vector<std::string> &args0)
@@ -1305,22 +442,23 @@ BlackBoxExec::BlackBoxExec(const std::string &program0,
 
 BlackBoxExec::~BlackBoxExec(void) {
   Support::Lock lock(mutex);
-  for (Session *s : sessions) {
+  for (BlackBoxProcessSession *s : sessions) {
     delete s;
   }
   sessions.clear();
 }
 
 
-BlackBoxExec::Session &BlackBoxExec::session(void) {
+BlackBoxProcessSession& BlackBoxExec::session(void) {
   Support::Lock lock(mutex);
-  for (Session *s : sessions) {
+  for (BlackBoxProcessSession *s : sessions) {
     if (s->owned_by_current_thread()) {
       return *s;
     }
   }
-  std::unique_ptr<Session> s(new Session(program, args));
-  Session *r = s.get();
+  std::unique_ptr<BlackBoxProcessSession> s(
+    create_blackbox_process(program, args));
+  BlackBoxProcessSession *r = s.get();
   sessions.push_back(r);
   s.release();
   return *r;
@@ -1468,7 +606,8 @@ decode_blackbox_response(const std::string& response, BlackBoxCall& call) {
 
 void
 BlackBoxExec::run(BlackBoxCall& call) {
-  const std::string response = session().run(encode_blackbox_request(call));
+  const std::string response =
+    session().exchange(encode_blackbox_request(call));
   decode_blackbox_response(response, call);
 }
